fix: show banner on main page
